Move classes to examscore package
Move associated test classes appropriately

Update import statements
