+++ src/main/java/seedu/address/logic/parser/SortCommandParser.java
@@ -6,7 +6,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 
 import seedu.address.logic.commands.SortCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Exam;
+import seedu.address.model.person.examscore.Exam;
 
 /**
  * Parses input arguments and creates a new SortCommand object.
