+++ src/main/java/seedu/address/model/AddressBook.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.util.ToStringBuilder;
-import seedu.address.model.person.Exam;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.UniquePersonList;
+import seedu.address.model.person.examscore.Exam;
 
 /**
  * Wraps all data at the address-book level
