+++ src/main/java/seedu/address/model/Model.java
@@ -5,8 +5,8 @@ import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.model.person.Exam;
 import seedu.address.model.person.Person;
+import seedu.address.model.person.examscore.Exam;
 
 /**
  * The API of the Model component.
