+++ src/main/java/seedu/address/model/ModelManager.java
@@ -11,8 +11,8 @@ import javafx.collections.ObservableList;
 import javafx.collections.transformation.FilteredList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.model.person.Exam;
 import seedu.address.model.person.Person;
+import seedu.address.model.person.examscore.Exam;
 
 /**
  * Represents the in-memory model of the address book data.
