+++ src/main/java/seedu/address/model/person/examscore/Exam.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.person.examscore;
 
 import static seedu.address.commons.util.AppUtil.checkArgument;
