+++ src/main/java/seedu/address/model/person/examscore/ExamScores.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.person.examscore;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
