+++ src/main/java/seedu/address/storage/JsonSerializableAddressBook.java
@@ -11,8 +11,8 @@ import com.fasterxml.jackson.annotation.JsonRootName;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.AddressBook;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.person.ExamList;
 import seedu.address.model.person.Person;
+import seedu.address.model.person.examscore.ExamList;
 
 /**
  * An Immutable AddressBook that is serializable to JSON format.
