+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -22,9 +22,9 @@ import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.person.Exam;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
+import seedu.address.model.person.examscore.Exam;
 import seedu.address.testutil.TestPersonBuilder;
 
 public class AddCommandTest {
