+++ src/test/java/seedu/address/logic/parser/EditScoreCommandParserTest.java
@@ -15,8 +15,8 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.EditScoreCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Exam;
-import seedu.address.model.person.ExamList;
+import seedu.address.model.person.examscore.Exam;
+import seedu.address.model.person.examscore.ExamList;
 
 public class EditScoreCommandParserTest {
