+++ src/test/java/seedu/address/logic/parser/SortCommandParserTest.java
@@ -6,14 +6,14 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_EXAM;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
-import static seedu.address.model.person.ExamList.FINAL;
-import static seedu.address.model.person.ExamList.MIDTERM;
+import static seedu.address.model.person.examscore.ExamList.FINAL;
+import static seedu.address.model.person.examscore.ExamList.MIDTERM;
 
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.SortCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.ExamList;
+import seedu.address.model.person.examscore.ExamList;
 
 /**
  * Contains unit tests for SortCommandParser
