+++ src/test/java/seedu/address/storage/JsonAdaptedPersonTest.java
@@ -15,12 +15,12 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.person.Attendance;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.ExamList;
-import seedu.address.model.person.ExamScores;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Phone;
 import seedu.address.model.person.StudentId;
 import seedu.address.model.person.TelegramHandle;
+import seedu.address.model.person.examscore.ExamList;
+import seedu.address.model.person.examscore.ExamScores;
 
 public class JsonAdaptedPersonTest {
     private static final String INVALID_NAME = "R@chel";
