+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -28,12 +28,12 @@ import java.util.Set;
 import seedu.address.model.AddressBook;
 import seedu.address.model.person.Attendance;
 import seedu.address.model.person.Email;
-import seedu.address.model.person.ExamList;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
 import seedu.address.model.person.StudentId;
 import seedu.address.model.person.TelegramHandle;
+import seedu.address.model.person.examscore.ExamList;
 import seedu.address.model.tag.Tag;
 
 /**
